refactor: :recycle: Re organize routes
