+++ __tests__/pages/Services/TcpTerminatedConnections.spec.tsx
@@ -3,8 +3,6 @@ import { Suspense } from 'react';
 import { render, waitForElementToBeRemoved } from '@testing-library/react';
 import { Server } from 'miragejs';
 
-
-import tcpConnections from '../../../mocks/data/TCP_CONNECTIONS.json';
 import servicesData from '../../../mocks/data/SERVICES.json';
 import { loadMockServer } from '../../../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../../../src/config/app';
@@ -14,7 +12,6 @@ import { Providers } from '../../../src/providers';
 import TcpTerminatedConnections from '../../../src/pages/Services/components/TcpTerminatedConnections';
 
 const servicesResults = servicesData.results;
-const tcpConnectionTerminated = tcpConnections.results[4];
 
 describe('Begin testing the TCP connections component', () => {
   let server: Server;
