--- src/pages/Services/routes.tsx
@@ -1,17 +1,0 @@
-import { lazy } from 'react';
-
-import { ServicesRoutesPaths } from './Services.enum';
-
-const Services = lazy(() => import(/* webpackChunkName: "services" */ './views/Services'));
-const Service = lazy(() => import(/* webpackChunkName: "service" */ './views/Service'));
-
-export const servicesRoutes = [
-  {
-    path: ServicesRoutesPaths.Services,
-    element: <Services />
-  },
-  {
-    path: `${ServicesRoutesPaths.Services}/:id`,
-    element: <Service />
-  }
-];
